user page 생성, 기존 media-list를 재활용하기 위해 media-search-query랑 분리
